requirements.txt
Flask~=3.0.3
gunicorn~=22.0.0
transformers~=4.40.1
torch~=2.3.0
sentencepiece~=0.2.0
tokenizers~=0.19.1
pymongo~=4.7.2
# pypdf2~=3.0.1
dateparser~=1.0.0
numpy~=1.26.4
scikit-learn~=0.24.2
textract~=1.6.5
gensim~=4.1.0
nltk~=3.8.1
textblob~=0.15.3
lxml~=5.2.2
# beautifulsoup4~=4.8.2
soupsieve~=2.5
dnspython~=2.6.1
# idna~=3.7
pillow~=10.3.0
olefile~=0.47
pip~=22.2.2
wheel~=0.37.1
jinja2~=3.1.4
regex~=2024.5.15
scipy~=1.10.1
click~=8.1.7
tqdm~=4.66.4
requests~=2.32.2
XlsxWriter~=3.2.0
pytz~=2024.1
colorama~=0.4.6
PyYAML~=6.0.1
MarkupSafe~=2.1.5
werkzeug~=3.0.3
blinker~=1.8.2
itsdangerous~=2.2.0
Cython~=0.29.23
setuptools~=65.3.0
mpmath~=1.3.0
sympy~=1.12
safetensors~=0.4.3
packaging~=24.0
urllib3~=2.2.1
fsspec~=2024.5.0
joblib~=1.4.2
python-dateutil~=2.9.0.post0
xlrd~=1.2.0
certifi~=2024.2.2
chardet~=3.0.4
# threadpoolctl~=3.5.0
tzlocal~=5.2
six~=1.12.0
docx2txt~=0.8
networkx~=3.2.1
argcomplete~=1.10.3
IMAPClient~=2.1.0
zipp~=3.18.2
wrapt~=1.16.0
ebcdic~=1.1.1
filelock~=3.14.0